% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_counts}
\alias{atac_counts}
\title{Example tibbles with input ATAC data in the format for (R)CONGAS+}
\format{
Example tibbles with input ATAC data.
}
\usage{
data(atac_counts)
}
\description{
Example tibbles with input data in the format for (R)CONGAS+
}
\examples{
data(atac_counts)
print(atac_counts)
}
\keyword{datasets}
