% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{auto_config_run}
\alias{auto_config_run}
\title{Title}
\usage{
auto_config_run(
  x,
  K,
  NB_size_atac = 150,
  NB_size_rna = 150,
  lambda = 0.3,
  a_sd = 1,
  b_sd = 100,
  prior_cn = c(0.2, 0.6, 0.2, 0.05, 0.025, 0.025),
  hidden_dim = length(prior_cn),
  init_importance = 0.6
)
}
\arguments{
\item{x}{}

\item{K}{}

\item{NB_size_atac}{}

\item{NB_size_rna}{}

\item{lambda}{}

\item{a_sd}{}

\item{b_sd}{}

\item{prior_cn}{}
}
\value{

}
\description{
Title
}
