% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{auto_normalisation_factor}
\alias{auto_normalisation_factor}
\title{Compute empirical normalisation factors (library size)}
\usage{
auto_normalisation_factor(x)
}
\arguments{
\item{x}{A tibble with a \code{cell} and \code{value} columns.}
}
\value{
A tibble, aggregated by cell with total values (sum), rescaled
by the estimated constant.
}
\description{
For a tibble dataset with a \code{cell} and \code{value}
columns, it sums up all values per cell (total library size), and scales
the value for \code{10^x} where \code{x} is the median nuimber of digits
in each total value size.

For instance for numbers of the order of ~1000 it will rescale the library
size by a factor 1000, taking all values around ~1.
}
\examples{
data('example_input')
example_input$x_rna \%>\% auto_normalisation_factor
}
