% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_xeno_10x}
\alias{breast_xeno_10x}
\title{Example Rcongas fit.}
\format{
Output from `run_inference`.
}
\usage{
data(breast_xeno_10x)
}
\description{
Data from ....
}
\examples{
data(breast_xeno_10x)
print(breast_xeno_10x)
plot(breast_xeno_10x)
}
\keyword{datasets}
