% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_clone_tree.R
\name{build_clone_tree}
\alias{build_clone_tree}
\title{Build Clone Tree form a CONGAS+ object Using medicc2}
\usage{
build_clone_tree(
  x,
  medicc_path = NULL,
  result_dir = "./medicc_results",
  input_dir = "./medicc_input",
  clean_inputs = TRUE,
  plot = TRUE
)
}
\arguments{
\item{x}{A CONGAS+ object with a valid fit.}

\item{medicc_path}{The file path of the \code{medicc2} executable. If not supplied, the function will attempt to find it in the system path.}

\item{result_dir}{The directory to save the results of medicc2. Defaults to "./medicc_results".}

\item{input_dir}{The directory where input files for medicc2 will be saved. Defaults to "./medicc_input".}

\item{clean_inputs}{Logical, should the function delete the input files after the execution of medicc2? Defaults to TRUE.}

\item{plot}{Logical, should the function plot the resulting tree? Defaults to TRUE.}
}
\value{
A list containing the newick tree (\code{tree}) and the plot (\code{plot}) if \code{plot} is set to TRUE.
}
\description{
This function builds a clone tree using the medicc2 tool. The function takes care of input file preparation from CONGAS* results,
running medicc2, cleaning up input files after execution, and plotting the tree using the \code{ggtree} package.
}
\examples{
\dontrun{
build_clone_tree(x, medicc_path = "/path/to/medicc2")
}


}
