% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{calculate_DE}
\alias{calculate_DE}
\title{Title}
\usage{
calculate_DE(
  X,
  input,
  clone1,
  clone2,
  method = "wilcox",
  normalize = TRUE,
  logfc.threshold = 0.25
)
}
\arguments{
\item{logfc.threshold}{}
}
\value{

}
\description{
Title
}
