% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campbell_bcpdx}
\alias{campbell_bcpdx}
\title{(R)CONGAS+ object for the breast cancer patient-derived xenograft,
first used in the original CONGAS paper.}
\format{
(R)CONGAS+ object.
}
\usage{
data(campbell_bcpdx)
}
\description{
Example object from (R)Congas+
}
\examples{
data(campbell_bcpdx)
print(campbell_bcpdx)
}
\keyword{datasets}
