% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_utilities.R
\name{cap_values_by_quantile}
\alias{cap_values_by_quantile}
\title{Cap observed values by quantile.}
\usage{
cap_values_by_quantile(x, upper_quantile = 0.98)
}
\arguments{
\item{x}{An input RNA/ATAC dataset where entries are indexeable by genomic
coordinate: "chr", "from" and "to".}

\item{upper_quantile}{The maximum quantile to determine cuts. If a value
is above the quantile it is capped.}
}
\value{
The input data with capped entries.
}
\description{
Cap observed values by quantile.
}
\examples{
data('example_input')
cap_values_by_quantile(example_input$x_rna, upper_quantile = .98)
cap_values_by_quantile(example_input$x_atac, upper_quantile = .98)
}
