% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{change_spot}
\alias{change_spot}
\title{Add a deletion or an insertion in a given segment}
\usage{
change_spot(parent_ploidy, mutation_spots, changes)
}
\arguments{
\item{parent_ploidy}{vector of integers, CNV value in the parent clone}

\item{mutation_spots}{vector of integers, index of the segmnts to be changed in the parent}

\item{changes}{list of integers, indicating for each ploidy status the allowed changes (e.g., list('1' = c(1,2)) since we cannot go to 0 copies)}
}
\value{
the vector with the new subclonal ploidy
}
\description{
Add a deletion or an insertion in a given segment
}
