% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congas_example}
\alias{congas_example}
\title{Example Rcongas fit.}
\format{
Output from `run_inference`.
}
\usage{
data(congas_example)
}
\description{
Data from triple negative breast cancer PDX. 
Provides an example of fitted Rcongas object
}
\examples{
data(congas_example)
print(congas_example)
plot(congas_example)
}
\keyword{datasets}
