% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tibble.r
\name{create_congas_tibble}
\alias{create_congas_tibble}
\title{Format the input counts before creating the CONGAS+ object}
\usage{
create_congas_tibble(
  counts,
  features = NULL,
  modality = "ATAC",
  save_dir = NULL,
  output_file = "counts_final.tsv"
)
}
\arguments{
\item{counts}{(Required). Count matrix with cells in columns and genes/bins in rows.}

\item{features}{(Required) Dataframe containing columns chr, from, to and (only if RNA) gene.}

\item{modality.}{Default 'ATAC' It has to be either ATAC or RNA based on the content of the count matrix.}

\item{output_dir.}{Optional. Default = NULL. Directory where to save the tibble as a \code{tsv} file. If NULL, the tibble object is only returned.}

\item{output_file.}{Optional, default is 'counts_final.tsv'. Name to give to the counts file. Note that this has not effect is \code{output_dir} is NULL.}
}
\description{
This function takes in input a SummarizedExperiment object with rowranges and  and creates a tibble with counts needed to create the Rcongas object.
}
