% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{create_df_from_range}
\alias{create_df_from_range}
\title{Create a dataframe from a breakpoint list}
\usage{
create_df_from_range(breaks, name)
}
\arguments{
\item{breaks}{vector of breakpoints for a given chromosome}

\item{name}{chhr name}
}
\value{
a data frame with genomic segments (with 3 cols chr, from, to)
}
\description{
Convert a chr and a breakpoint vector in a segments data frame
}
