% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{divide_genome_in_segments}
\alias{divide_genome_in_segments}
\title{A simple function to divide the genome in segments}
\usage{
divide_genome_in_segments(approx_length = 1e+08)
}
\arguments{
\item{approx_length}{Approximate segment legnth. Integer}
}
\value{
A data frame with genomic segments (with 3 cols chr, from, to)
}
\description{
The functions first finds the maximum number of segments of a given length for each chromosome in the
human genome (hg38 assembly, onlys chrs 1:22, X and Y) then it sample the number of actual splits
in a range (0,MAX). In the end it adds some stochasticity to the segment ends and starts.
}
