% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_input}
\alias{example_input}
\title{Example tibbles with input data in the format for (R)Congas+}
\format{
Example tibbles with input data.
}
\usage{
data(example_input)
}
\description{
Example tibbles with input data in the format for (R)Congas+
}
\examples{
data(example_input)
print(example_input)
}
\keyword{datasets}
