% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_utilities.R
\name{filter_cells_by_quantile}
\alias{filter_cells_by_quantile}
\title{Filter cells by upper and lower quantile of their normalisation factors.}
\usage{
filter_cells_by_quantile(x, lower_quantile = 0.05, upper_quantile = 0.95)
}
\arguments{
\item{x}{A tibble with a cell and normalisation_factor column.}

\item{lower_quantile}{The minimum quantile to determine cuts.}

\item{upper_quantile}{The maximum quantile to determine cuts.}
}
\value{
The input data with normalisation factors in between the quantiles range.
@export
}
\description{
Filter cells by upper and lower quantile of their normalisation factors.
}
\examples{
data('example_input')
}
