% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clusters.R
\name{filter_clusters}
\alias{filter_clusters}
\title{Filter output clusters}
\usage{
filter_clusters(x, ncells = 10, abundance = 0.03)
}
\arguments{
\item{x}{Rcongas object}

\item{ncells}{minimum size of a valid cluster expressed as absolute number of cells}

\item{abundance}{minimum size of a cluster expressed as a percentage of the total cell number}
}
\value{

}
\description{
From a CONGAS fit object remove clusters that do not pass filters.
These include number of cells and abundance of cluster (mixing proportions). This function the assigns the 
cells back to the most similar remaining cluster. In case the inference was performed using MAP we choose the cluster with
the most similar CNV profile (based on euclidean distance), otherwise we assign the cell to the cluster with second highest 
probability and renormalize the z.
}
\examples{

x = Rcongas::congas_example

print(x)

# Equivalent filters for this model
x \%>\% Rcongas:::filter_clusters(abundance = .5)

x \%>\% Rcongas:::filter_clusters(ncells = 150)


}
