% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clusters_and_remove_cells.R
\name{filter_clusters_and_remove_cells}
\alias{filter_clusters_and_remove_cells}
\title{Filter output clusters and remove cells beloging to those clusters}
\usage{
filter_clusters_and_remove_cells(x, ncells = 10, abundance = 0.03)
}
\arguments{
\item{x}{Rcongas object}

\item{ncells}{minimum size of a valid cluster expressed as absolute number of cells}

\item{abundance}{minimum size of a cluster expressed as a percentage of the total cell number}
}
\value{

}
\description{
The function works in a similar way to \code{\link{filter_clusters}} but it also removes the cells belonging to those clusters
}
\examples{




}
