% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_counts_by_quantile}
\alias{filter_counts_by_quantile}
\title{Filter observed counts by quantile.}
\usage{
filter_counts_by_quantile(x, upper_quantile = 0.98)
}
\arguments{
\item{x}{An input RNA/ATAC dataset where entries are indexeable by genomic
coordinate: "chr", "from" and "to".}

\item{upper_quantile}{The maximum quantile to determine cuts. If a value
is above the quantile it is removed}
}
\value{
The input data with removed entries.
}
\description{
Filter observed counts by quantile.
}
\examples{
data('example_input')
filter_counts_by_quantile(example_input$rna, upper_quantile = .98)
filter_counts_by_quantile(example_input$atac, upper_quantile = .98)
}
