% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter_genes}
\alias{filter_genes}
\title{Removing highly variable and highly expressed genes}
\usage{
filter_genes(cmat, quantile_mean = 0.95, quantile_var = 0.95, plot = TRUE)
}
\arguments{
\item{cmat}{}

\item{quantile_mean}{}

\item{quantile_var}{}

\item{plot}{}
}
\value{

}
\description{
Removing highly variable and highly expressed genes
}
\examples{

}
