% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_utilities.R
\name{filter_missing_data}
\alias{filter_missing_data}
\title{Filter cells with missing data.}
\usage{
filter_missing_data(x, proportion_RNA = 0.05, proportion_ATAC = 0.05)
}
\arguments{
\item{x}{An \code{rcongasplus} object.}

\item{proportion_RNA}{The RNA proportion cut for a cell to be removed, default 5\%.}

\item{proportion_ATAC}{The ATAC proportion cut for a cell to be removed, default 5\%.}
}
\value{
The object \code{x} where 0-counts cells have been removed.
}
\description{
After mapping counts data to segments, this function
can be used to determine cells with missing data, and remove
them The function requires and returns an (R)CONGAS+ object.

This filter works by a proportion, as reported by the \code{\link{stat}} function.

If these cells are not removed, during inference missing values
are imputed to be \code{0}. This can create an excess of mixture components
fitting 0-counts data.
}
\examples{
data('example_object')

# Default
print(example_object)

example_object = filter_missing_data(example_object)

# After filtering
print(example_object)
}
