% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_utilities.R
\name{filter_outliers}
\alias{filter_outliers}
\title{Filter per-segment outliers by quantiles.}
\usage{
filter_outliers(
  x,
  lower_quantile = 0.01,
  upper_quantile = 0.99,
  action = "cap"
)
}
\arguments{
\item{x}{An \code{rcongasplus} object.}

\item{lower_quantile}{The lower quantile, default 1\%.}

\item{upper_quantile}{The upper quantile, default 99\%.}

\item{action}{If \code{"remove"}, outliers will be set to 0. If \code{"cap"},
outliers will be capped at the median per-cell counts.}
}
\value{
The object \code{x} where outliers have been identified and removec
or capped according to the parameters.
}
\description{
After mapping counts data to segments, this function
can be used to determine quantiles of mapped data, and identify
outliers in each segment and modality.

An outlier can then be removed or capped to the median cell value.
The former option introduced 0-counts in the data, which we suggest
to check with the \code{stat} function, and possibly remove by using
the \code{filter_missing_data} function. Removal can be important as
an excess of 0-counts cells (\emph{missing data}) will drive the fit
to use 0-mean components.

Capping does not introduce any 0-count cell, and is the suggested choice.
The capped values is either a count value or a z-score, depending on
the modality type of likelihood.

In both cases pre-filtering normalisation factors are no longer adequate
after filtering, and have to be recomputed. If the modality adopts a
Gaussian likelihood this is not a problem, since those are set to 1
when the object is created, and remain 1 afterwards. In the case of counts
based likelihood like Negative Binomials these are re-computed for all
input cells by using the \code{auto_normalisation_factor} function.

Therefore, if custom factors have been computing this function might
affect the general signal in the data, and factors should be handled
explicitly by the user.

The function requires and returns an (R)CONGAS+ object.
}
\examples{
data('example_object')

# Default
print(example_object)

example_object \%>\% 
  filter_outliers(, action = 'cap') \%>\% 
  print()

example_object \%>\% 
  filter_outliers(, action = 'remove') \%>\% 
  print()
}
