% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_quantiles_per_sample.R
\name{filter_quantiles_per_segments}
\alias{filter_quantiles_per_segments}
\title{Per segment cut of high and lower quantiles}
\usage{
filter_quantiles_per_segments(x, low_quant = 0.01, high_quant = 0.99)
}
\arguments{
\item{x}{Rcongas object}

\item{low_quant}{lower quantile}

\item{high_quant}{high quantile}
}
\value{

}
\description{
The function simply calculates two quantiles for each segments and the caps all the values higher than `high_quant` and lower than 
`low_quant`
}
\examples{

x <-  Rcongas::congas_example

get_counts(x) \%>\% pull(n) \%>\% max

x <- filter_quantiles_per_segments(x)

get_counts(x) \%>\% pull(n) \%>\% max

}
