% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mapped.R
\name{filter_segments}
\alias{filter_segments}
\title{Filter segments.}
\usage{
filter_segments(
  x,
  length = 0,
  RNA_genes = 50,
  ATAC_peaks = 50,
  segments = list(),
  RNA_nonzerocells = 0,
  ATAC_nonzerocells = 0
)
}
\arguments{
\item{x}{An \code{rcongasplus} object.}

\item{length}{Mimium size in number of bases for each segment. This is \code{0}
by default, and therefore non effective.}

\item{RNA_genes}{Required number of RNA genes, if RNA is available. This is
\code{50} by default.}

\item{ATAC_peaks}{Required number of ATAC peaks, if ATAC is available. This is
\code{50} by default.}

\item{segments}{A list of segment ids to retain regardless of the filters.}
}
\value{
The object \code{x} where segments have been identified and
removed.
}
\description{
After mapping data to segments, this function
can be used to remove segments that are too short, or that have too
few mapped RNA genes or ATAC peaks. A parameters allows to input a list
of segments that will be retained no matter what.

The function requires and returns an (R)CONGAS+ object.
}
\examples{
data('example_object')

# Default
print(example_object)

example_object \%>\% 
  filter_segments() \%>\% 
  print()

# Minimum size
example_object \%>\% 
  filter_segments(length = 1e7) \%>\% 
  print()
}
