% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_utilities.R
\name{filter_values_by_quantile}
\alias{filter_values_by_quantile}
\title{Filter observed values by quantile.}
\usage{
filter_values_by_quantile(x, upper_quantile = 0.98)
}
\arguments{
\item{x}{An input RNA/ATAC dataset where entries are indexeable by genomic 
coordinate: "chr", "from" and "to".}

\item{upper_quantile}{The maximum quantile to determine cuts. If a value
is above the quantile it is removed}
}
\value{
The input data with removed entries.
}
\description{
Filter observed values by quantile.
}
\examples{
data('example_input')
filter_values_by_quantile(example_input$x_rna, upper_quantile = .98)
filter_values_by_quantile(example_input$x_atac, upper_quantile = .98)
}
