% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate_cluster_ploidy_df}
\alias{generate_cluster_ploidy_df}
\title{generate a dataframe with segments coordinates and corresponding ploidy values for the specified number of clusters.}
\usage{
generate_cluster_ploidy_df(
  K = 2,
  spots = 5,
  probs = c(0, 1, 0, 0),
  div_factor_dist = 1e+06,
  random = TRUE,
  changes = list(`0` = c(1, 2), `1` = c(1), `2` = c(-1, 1), `3` = c(-1, 1), `4` = c(-1,
    1), `5` = c(-1), `6` = c(-1)),
  fine = FALSE,
  karyo = 1:4,
  type = "parent",
  non_random_spots = c(17, 18),
  spots_deletion = c(17, 18),
  spots_insertions = c(8),
  spots_O = c(),
  tree = generate_random_tree(K),
  segments = divide_genome_in_segments(),
  baf = TRUE
)
}
\arguments{
\item{K}{integer, number of clusters to simulate}

\item{spots}{number of segment to change when using random mode}

\item{probs}{probabilities of picking a given CN value from the karyotype, needs to have length = length(karyo)}

\item{random}{shold the mutated segments among parent-child subclones be choosen at random?}

\item{changes}{what are th allowed changes in ploidy between parent-child subclnes when fine = FALSE}

\item{fine}{allow a finer controll on the different segments}

\item{karyo}{integer vector, alloed CN values}

\item{type}{one of c("parent", "random"), indicates if the clusters have an evolutionary relionship or they are unrelated}

\item{non_random_spots}{if random = FALSE, wihch spotss should harbor a mutation?}

\item{spots_deletion}{segments where we want to insert a single chromosome deletion}

\item{spots_O}{segments where we wanto to insert a double chromosome deletion}

\item{tree}{a data frame with two columns, the first with the index of each cluster and the second with its parent}

\item{spots_insertion}{segments where we want to insert a single or double (randomly sampled)
chromose amplification}

\item{approx_length}{approximate segment length}
}
\value{
df with segments and ploidy information
}
\description{
generate a dataframe with segments coordinates and corresponding ploidy values for the specified number of clusters.
}
