% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate_ploidy}
\alias{generate_ploidy}
\title{Generete the ploidy for a segment}
\usage{
generate_ploidy(length, probs, karyo)
}
\arguments{
\item{length}{integer, length of the sample (basically it should be always one)}

\item{probs}{vector with the same length as karyo, the probability of picking a given CNV value}

\item{karyo}{vector of integer, CNV states allowed}
}
\value{
an integer, the CNV state for tthat segment
}
\description{
Generete the ploidy for a segment
}
