% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate_ploidy_parent}
\alias{generate_ploidy_parent}
\title{Generat the ploidy given a parent clone}
\usage{
generate_ploidy_parent(parent_ploidy, spots, random, non_random_spots, changes)
}
\arguments{
\item{spots}{integer, how many sites to be changed}

\item{random}{boolean, should the spots be sampled randomly}

\item{non_random_spots}{vector of integers,}

\item{changes}{vector of integers, allowed values for deletion and amplification}

\item{paret_ploidy}{vector of integers, the ploidy of the parent clone}
}
\value{
vector of integers, ploidy profile of the child subclone
}
\description{
Generat the ploidy given a parent clone
}
