% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate_ploidy_parent_fine}
\alias{generate_ploidy_parent_fine}
\title{Generate the ploidy given a parent clone (with a finer control)}
\usage{
generate_ploidy_parent_fine(
  parent_ploidy,
  spots_deletion,
  spots_insertions,
  spots_O
)
}
\arguments{
\item{spots_deletion}{segments where we want to insert a single chromosome deletion}

\item{spots_O}{segments where we wanto to insert a double chromosome deletion}

\item{spots_insertion}{segments where we want to insert a single or double (randomly sampled)
chromose amplification}
}
\value{
vector of integers, ploidy profile of the child subclone
}
\description{
It is basically equal to {\link{generate_ploidy_parent}} but provides the direct controll of the segments
in which we want to place the mutation
}
