% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_random_tree}
\alias{generate_random_tree}
\title{Generate a random tree}
\usage{
generate_random_tree(K, unrelated_prob = 0)
}
\arguments{
\item{K}{number of subclones in the population}

\item{unrelated_prob}{probability of the emergence of a separate population}
}
\value{
a data frae with two columns indicating the cluster and its parent cluster
}
\description{
Generate a random tree
}
