% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Extract cell clustering assignments.}
\usage{
get_clusters(x, clusters = NULL, cut_znk = 0)
}
\arguments{
\item{x}{Input object with clusters.}

\item{cut_znk}{Cutoff for latent variables.}

\item{cluster_label}{Cluster id to subset the assignments.}
}
\value{
A tibble with the cell id, cluster labels, latent variables values
for every clustering id, and \code{p_assignment} the hard-clustering assignment
probability for the cells (max among the latent variables).
}
\description{
Extract a tibble for clustering assignments of the input cells,
possibly subsetting by cluster or leaving \code{NA} those cell assignments
that have probability ($z_{nk}$) below a cutoff (via latent variables).
}
\examples{

x = Rcongas::congas_example

# Default view
get_clusters(x)

# Assignments to specific cluster
get_clusters(x, clusters = "c1")

# Set to NA all clustering assignments with p_assignment below a 99\%
get_clusters(x, cut_znk = .99)
}
