% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters_size.R
\name{get_clusters_size}
\alias{get_clusters_size}
\title{Extract size of clusters.}
\usage{
get_clusters_size(x, normalised = FALSE)
}
\arguments{
\item{x}{Input object with clusters.}

\item{normalised}{If \code{TRUE}, normalises the proportions.}
}
\value{
A named vector for cluster sizes.
}
\description{
Compute the size of CONGAS clusters, normalised or in absolute
cell numbers.
}
\examples{

x = Rcongas::congas_example

get_clusters_size(x, normalised = FALSE)
get_clusters_size(x, normalised = TRUE)
}
