% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_input_data.R
\name{get_counts}
\alias{get_counts}
\title{Title}
\usage{
get_counts(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  normalise = TRUE,
  z_score = FALSE,
  sum_denominator = TRUE
)
}
\arguments{
\item{normalise}{}
}
\value{

}
\description{
Title
}
