% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_stats.R
\name{get_dataset_stats}
\alias{get_dataset_stats}
\title{Return dataset statistics}
\usage{
get_dataset_stats(x)
}
\arguments{
\item{x}{A CONGAS object}
}
\value{
A list of named values that contain the required statistics.
}
\description{
From a CONGAS object returns the data statistics such ash
numnber of cells, semgents, clusters (if any), score used etc.
}
\examples{

x = Rcongas::congas_example

get_dataset_stats(x)
}
