% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_extra_analysis.R
\name{get_DE_table}
\alias{get_DE_table}
\title{Get a table of DE genes}
\usage{
get_DE_table(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  cut_pvalue = 0.001,
  cut_lfc = 0.25
)
}
\arguments{
\item{x}{an Rcongas object after DE analysis}

\item{chromosomes}{filter for specific chromosomes}

\item{cut_pvalue}{corrected (BH) p-value filter}

\item{cut_lfc}{log fold change filter}
}
\value{
a table with differentially expressed genes
}
\description{
The actual function for DE is implemented following 
the Seurat `FindMarkers` method and as such conserves the same output structure
}
