% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_annotations.R
\name{get_gene_annotations}
\alias{get_gene_annotations}
\title{Genome coordinates available in the package.}
\usage{
get_gene_annotations(x)
}
\arguments{
\item{x}{An input CONGAS object.}
}
\value{
The required tibble.
}
\description{
From a CONGAS object, it teturns the  
tibbles for the genomic coordinates available in the package
that have been used to assemble the object.
}
\examples{

x = Rcongas::congas_example

get_gene_annotations(x)
}
