% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_raw_data.R
\name{get_input_raw_data}
\alias{get_input_raw_data}
\title{Title}
\usage{
get_input_raw_data(
  x,
  all_cells = FALSE,
  all_genes = FALSE,
  add_locations = FALSE,
  add_clusters = FALSE,
  as_matrix = FALSE
)
}
\arguments{
\item{all_cells}{}

\item{all_genes}{}

\item{as_matrix}{}
}
\value{

}
\description{
Title
}
