% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_genome.R
\name{get_reference_genome}
\alias{get_reference_genome}
\title{Reference genome used in a model.}
\usage{
get_reference_genome(x)
}
\arguments{
\item{x}{Input CONGAS model.}
}
\value{
A code for any of the supported references
}
\description{
Extracts the code for the reference genome used in a model.
}
\examples{

x = Rcongas::congas_example

get_reference_genome(x)
}
