% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segment_ids.R
\name{get_segment_ids}
\alias{get_segment_ids}
\title{Extract the segment ids (for indexing).}
\usage{
get_segment_ids(x, highlight = FALSE, ...)
}
\arguments{
\item{x}{Input object (with or without clustering).}

\item{highlight}{If \code{TRUE}, returns return only the segments that are
highlighted in the plots. Use ellipsis to specify parameters such as \code{alpha}
to function \code{get_clusters_ploidy}.}

\item{...}{Parameters forwarded to \code{get_clusters_ploidy} to get segment
ids.}
}
\value{
A vector of strings.
}
\description{
For the functions that require a segment identifier as input,
for instance \code{plot_segment_density},
this function returns the list of segment ids used in the input object. If 
run with \code{highlight = TRUE} it will return only the segments that
are highlighted in the plots; this requires clustering to be computed
}
\examples{

x = Rcongas::congas_example

# All segments
get_segment_ids(x)

# All segments that are highlighted with default value
get_segment_ids(x, highlight = TRUE)

# All segments that are highlighted with stricter value than default
get_segment_ids(x, highlight = TRUE, alpha = 0.01)
}
