% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_clustering_results.R
\name{get_segment_test_counts}
\alias{get_segment_test_counts}
\title{Title}
\usage{
get_segment_test_counts(x, group1, group2, cutoff_p = 0.01)
}
\arguments{
\item{cutoff_p}{}
}
\value{

}
\description{
Title
}
