% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_karyo}
\alias{hg38_karyo}
\title{Karyotype hg38}
\format{
dataframe
}
\usage{
data(hg38_karyo)
}
\description{
Data about the coordinates of hg38
}
\examples{
data(hg38_karyo)
print(hg38_karyo)
}
\keyword{datasets}
