% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List models available in the package.}
\usage{
list_models()
}
\value{
Nothing
}
\description{
Returns a simple description of the models that can be 
choosen to fit the data.
}
\examples{
list_models()
}
