% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Dataframe containing cell type annotations for the 10x multiome lymphoma dataset.}
\format{
Dataframe.
}
\usage{
data(metadata)
}
\description{
Dataframe with cell type annotations for the 10x multiome lymphoma dataset.
}
\examples{
data(metadata)
head(metadata)
}
\keyword{datasets}
