% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiome_congas_object}
\alias{multiome_congas_object}
\title{Multiome congas object}
\format{
CONGAS+ object
}
\usage{
data(multiome_congas_object)
}
\description{
Example of a CONGAS+ object containg a subset of the multiome Lymphoma data.
}
\examples{
data(multiome_congas_object)
print(multiome_congas_object)
}
\keyword{datasets}
