% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongas.R
\name{plot.rcongas}
\alias{plot.rcongas}
\title{Plot an Rcongas fit.}
\usage{
\method{plot}{rcongas}(x, ...)
}
\arguments{
\item{...}{}
}
\value{
A ggplot object for the plot.
}
\description{
Plot an Rcongas fit.
}
\examples{
x=1
}
