% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongasplus.R
\name{plot.rcongasplus}
\alias{plot.rcongasplus}
\title{Plot for an object of class \code{'rcongasplus'}.}
\usage{
\method{plot}{rcongasplus}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'rcongasplus'}.}

\item{...}{Unused ellipsis.}
}
\value{
One of the plots available in the package, computed with
\code{\link{plot_data}} or \code{\link{plot_data}} functions.
}
\description{
Plot for an object of class \code{'rcongasplus'}.
}
\examples{
data("example_object")
plot(example_object)
}
