% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_statistics.R
\name{plot_cell_statistics}
\alias{plot_cell_statistics}
\title{Title}
\usage{
plot_cell_statistics(x, cutoff_zeroes = 0.9, cutoff_counts = 0, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
