% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chromosome.R
\name{plot_chromosome}
\alias{plot_chromosome}
\title{Title}
\usage{
plot_chromosome(x, chr = "chr1", plot = c("segments", "counts", "DE"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

# Shows all data available
plot_chromosome(x)

# Chromosome 17
plot_chromosome(x, chr = 'chr17')
}
