% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chromosome_inspection.R
\name{plot_chromosome_inspection}
\alias{plot_chromosome_inspection}
\title{Title}
\usage{
plot_chromosome_inspection(
  x,
  chr,
  normalise = "ls",
  top = 5,
  n_sorrounding = 5,
  ...
)
}
\arguments{
\item{n_sorrounding}{}
}
\value{

}
\description{
Title
}
