% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cohort_statistics.R
\name{plot_cohort_statistics}
\alias{plot_cohort_statistics}
\title{Title}
\usage{
plot_cohort_statistics(x, cutoff = 0.9, assembly = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
