% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot data.}
\usage{
plot_data(x, what = "histogram", ...)
}
\arguments{
\item{what}{Any of \code{"histogram"}, \code{"lineplot"}, \code{"heatmap"} or
\code{"mapping"}.}

\item{segments}{Can subset the plot only to certain segments, which are indexed
by the \code{segment_id} format in the tibble returned by function \code{get_input}
with \code{what = 'segmentation'} as parameter.}
}
\value{
A \code{ggplot} or \code{cowplot} figure, depending on the number of
modalities and plot required.
}
\description{
General plot for data, showing data as a histogram, a lineplot
or a heatmap. Discrete count data are normalised by input factors.
}
