% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DE_volcano.R
\name{plot_DE_volcano}
\alias{plot_DE_volcano}
\title{Title}
\usage{
plot_DE_volcano(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  annotate_top = 5,
  cut_pvalue = 0.001,
  cut_lfc = 0.25
)
}
\arguments{
\item{cut_lfc}{}
}
\value{

}
\description{
Title
}
