% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_counts_on_chromosome.R
\name{plot_gene_counts_on_genome}
\alias{plot_gene_counts_on_genome}
\title{plot_gene_counts_on_genome}
\usage{
plot_gene_counts_on_genome(
  x,
  cutoff_quantile_sd = 0.95,
  normalise = c("ls", "cn"),
  ...
)
}
\arguments{
\item{x}{Input object.}

\item{cutoff_quantile_sd}{Shows only counts with standard deviation above this quantile (default >95%); 
it should avoid plotting a number of low-frequency data points.}

\item{normalise}{Normalise by library size (\code{"ls"}) and/or by input copy number (\code{"ls"}).}

\item{...}{Parameters forwarded to \code{get_genes_mapped_to_segments} and \code{get_input_raw_data}.}
}
\value{
A ggplot figure with multiple plots.
}
\description{
Plots gene per cell counts on the genome.
}
\examples{

plot_gene_counts_on_genome(Rcongas::congas_example)
}
