% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_rank.R
\name{plot_gene_rank}
\alias{plot_gene_rank}
\title{Title}
\usage{
plot_gene_rank(
  x,
  fun = sd,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  top = 300,
  by_segment = FALSE,
  assembly = TRUE,
  ...
)
}
\arguments{
\item{by_segment}{}
}
\value{

}
\description{
Title
}
