% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_stdev.R
\name{plot_gene_stdev}
\alias{plot_gene_stdev}
\title{Title}
\usage{
plot_gene_stdev(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  top = 300,
  by_segment = FALSE,
  ...
)
}
\arguments{
\item{by_segment}{}
}
\value{

}
\description{
Title
}
