% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gw_cna_profiles.R
\name{plot_gw_cna_profiles}
\alias{plot_gw_cna_profiles}
\title{Title}
\usage{
plot_gw_cna_profiles(
  x,
  whole_genome = FALSE,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  cutoff_p = 0.01
)
}
\arguments{
\item{chromosomes}{}
}
\value{

}
\description{
Title
}
