% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_highlight.R
\name{plot_highlights}
\alias{plot_highlights}
\title{Title}
\usage{
plot_highlights(x, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Title
}
\examples{

x = Rcongas::congas_example

# Default view 
plot_highlights(x) 

# More stringent 
plot_highlights(x, alpha = 0.01) 

# Even more stringent 
plot_highlights(x, alpha = 0.001)  
}
