% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inference_report.R
\name{plot_inference_report}
\alias{plot_inference_report}
\title{Title}
\usage{
plot_inference_report(x, assembly = TRUE)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
\examples{

# Report
plot_inference_report(x)
}
