% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_variables.R
\name{plot_latent_variables}
\alias{plot_latent_variables}
\title{Title}
\usage{
plot_latent_variables(x, cutoff_assignment = 0)
}
\arguments{
\item{cutoff_assignment}{}
}
\value{

}
\description{
Title
}
