% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixing_proportions.R
\name{plot_mixing_proportions}
\alias{plot_mixing_proportions}
\title{Title}
\usage{
plot_mixing_proportions(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
\examples{
x = Rcongas::congas_example

plot_mixing_proportions(x)
}
