% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normalisation_factors.R
\name{plot_normalization_factors}
\alias{plot_normalization_factors}
\title{Title}
\usage{
plot_normalization_factors(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
\examples{

x = Rcongas::congas_example

plot_normalization_factors(x) 
}
