% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{plot_raw_data}
\alias{plot_raw_data}
\title{Title}
\usage{
plot_raw_data(
  x,
  genes,
  description = "My CONGAS model",
  clusters = NULL,
  prompt = TRUE,
  ...
)
}
\arguments{
\item{normalised}{}
}
\value{

}
\description{
Title
}
\examples{

  x = get_input_raw_data(rcongas_example)
  genes = get_DE_table(rcongas_example) \%>\% pull(gene)
  clusters = get_clusters(rcongas_example)
  description = get_model_description(rcongas_example)
}
