% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report_segments.R
\name{plot_report_segments}
\alias{plot_report_segments}
\title{Title}
\usage{
plot_report_segments(x, chromosomes = paste0("chr", c(1:22, "X", "Y")), ...)
}
\arguments{
\item{chromosomes}{}
}
\value{

}
\description{
Title
}
