% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segment_density.R
\name{plot_segment_density}
\alias{plot_segment_density}
\title{Title}
\usage{
plot_segment_density(
  x,
  segments_ids,
  group1 = 1,
  group2 = 2,
  sum_denominator = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
