% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongas.R
\name{print.rcongas}
\alias{print.rcongas}
\title{Summaries for an object of class \code{'rcongas'} is like a print.}
\usage{
\method{print}{rcongas}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'rcongas'}.}

\item{...}{}
}
\value{
nothing.
}
\description{
Summaries for an object of class \code{'rcongas'} is like a print.
}
\examples{
data(fit_example)
print(fit_example$best)
}
