% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{randomize_breaks}
\alias{randomize_breaks}
\title{Add stocasticity in the brakpoints}
\usage{
randomize_breaks(breaks, chr, length, chr_lengths, shuffle_perc = 0.1)
}
\description{
Sample the noise from a uniform distribution \link{-(seg_length * shuffle_perc), (seg_length * shuffle_perc)}
and add it to the breakpoint
}
