% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcongas_example}
\alias{rcongas_example}
\title{Example Rcongas fit.}
\format{
Output from `run_inference`.
}
\usage{
data(rcongas_example)
}
\description{
Data from ....
}
\examples{
data(rcongas_example)
print(rcongas_example)
plot(rcongas_example)
}
\keyword{datasets}
