% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_information_criteria.R
\name{recalculate_information_criteria}
\alias{recalculate_information_criteria}
\title{Title}
\usage{
recalculate_information_criteria(x, method, normalize_by_segs)
}
\arguments{
\item{x}{}

\item{method}{}
}
\value{

}
\description{
Title
}
\examples{

}
