% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis.R
\name{report_analysis}
\alias{report_analysis}
\title{Report for a template set of results after CONGAS}
\usage{
report_analysis(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Report for a template set of results after CONGAS
}
\examples{

x = Rcongas::congas_example

print(x)

}
