% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_chr.R
\name{report_analysis_chr}
\alias{report_analysis_chr}
\title{Title}
\usage{
report_analysis_chr(
  x,
  normalise = "ls",
  top = 5,
  n_sorrounding = 5,
  cex = 1,
  ...
)
}
\arguments{
\item{cex}{}
}
\value{

}
\description{
Title
}
