% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_counts}
\alias{rna_counts}
\title{Example tibbles with input RNA data in the format for (R)CONGAS+}
\format{
Example tibble with input RNA data.
}
\usage{
data(rna_counts)
}
\description{
Example tibbles with input data in the format for (R)CONGAS+
}
\examples{
data(rna_counts)
print(rna_counts)
}
\keyword{datasets}
