% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{run_inference}
\alias{run_inference}
\title{Title}
\usage{
run_inference(
  X,
  model,
  optim = "ClippedAdam",
  elbo = "TraceGraph_ELBO",
  inf_type = "SVI",
  steps = 300,
  lr = 0.05,
  param_list = list(),
  MAP = TRUE,
  seed = 3,
  rerun = F,
  norm_factor = NULL
)
}
\arguments{
\item{step_post}{}
}
\value{

}
\description{
Title
}
