% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{satpathy_bcc}
\alias{satpathy_bcc}
\title{Basal cell carcinoma (10x RNA and ATAC, counts-based)}
\format{
(R)CONGAS+ object.
}
\usage{
data(satpathy_bcc)
}
\description{
(R)CONGAS+ object for the basal cell carcinoma dataset,
used in the CONGAS+ paper. The paper providing these data
is Satpathy, Ansuman T., et al. "Massively parallel single-cell chromatin
landscapes of human immune cell development and intratumoral T cell
exhaustion." Nature Biotechnology 37.8 (2019): 925-936.
}
\examples{
data(satpathy_bcc)
print(satpathy_bcc)
}
\keyword{datasets}
