% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_selector.R
\name{segments_selector_congas}
\alias{segments_selector_congas}
\title{Filter segments where the optimal number of clusters is 1.}
\usage{
segments_selector_congas(
  obj,
  K_max = 3,
  score = "BIC",
  lambda = 0.5,
  cores_ratio = 0.5
)
}
\arguments{
\item{obj}{An \code{rcongasplus} object.}

\item{K_max}{Max Number of clusters to test for the congas runs on single segments}

\item{score}{Model selection score to be used to select the number of clusters}

\item{lambda}{lambda to use for each congas run}

\item{cores_ratio}{fraction of cores that we be used to run the single CONGAS+ inferences in parallel.}
}
\value{
The object \code{obj} where segments have been identified and
removed.
}
\description{
This function can be used to exclude from the inference step those segments that are unimodal.

The function requires and returns an (R)CONGAS+ object.
}
