% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongas.R
\name{summary.rcongas}
\alias{summary.rcongas}
\title{Summary for an object of class \code{'rcongas'}.}
\usage{
\method{summary}{rcongas}(object, ...)
}
\arguments{
\item{...}{Extra parameters.}

\item{x}{Object of class \code{'rcongas'}.}
}
\value{
See \code{\link{print}}.
}
\description{
The summary is equivalent to a \code{print}.
}
\examples{
data(fit_example)
summary(fit_example$best)
}
