% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zaho_hemato}
\alias{zaho_hemato}
\title{Hematopoietic bone marrow failure (10x RNA, counts-based)}
\format{
(R)CONGAS+ object.
}
\usage{
data(zaho_hemato)
}
\description{
(R)CONGAS+ object for the hematopoietic dataset,
first used in the original CONGAS paper. The paper providing these data
is Zhao, Xin, et al. "Single-cell RNA-seq reveals a distinct transcriptome
signature of aneuploid hematopoietic cells." Blood 130.25 (2017): 2762-2773.
}
\examples{
data(zaho_hemato)
print(zaho_hemato)
}
\keyword{datasets}
