% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_theta_prior_params.R
\name{approx_theta_prior_params}
\alias{approx_theta_prior_params}
\title{Approximation of the library size prior}
\usage{
approx_theta_prior_params(x, plot = FALSE, mult_factor = 1, bins = 30)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
This function fit a theta
}
