% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R
\name{auto_config_run}
\alias{auto_config_run}
\title{Determine data-based hyperparameters for the Bayesian priors.}
\usage{
auto_config_run(
  x,
  K,
  NB_size_atac = 150,
  NB_size_rna = 150,
  lambda = 0.3,
  a_sd = 1,
  b_sd = 100,
  prior_cn = c(0.2, 0.6, 0.2, 0.05, 0.025, 0.025),
  hidden_dim = length(prior_cn),
  init_importance = 0.6,
  NB_size_priors = c(15, 1000)
)
}
\arguments{
\item{prior_cn}{}
}
\value{

}
\description{

}
