% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cluster.R
\name{best_cluster}
\alias{best_cluster}
\title{Model selection with CONGAS.}
\usage{
best_cluster(
  X,
  model,
  clusters,
  optim = "ClippedAdam",
  elbo = "TraceEnum_ELBO",
  inf_type = "SVI",
  steps = 300,
  lr = 0.05,
  param_list = list(),
  MAP = TRUE,
  seed = 3,
  mixture = NULL,
  method = "AIC",
  normalize_by_segs = FALSE
)
}
\arguments{
\item{X}{rcongas object generated by the constructor \fun{init}}

\item{model}{Any of the model keywords available via \code{list_models()}.}

\item{clusters}{A vector of values for \code{k}.}

\item{optim}{a pyro.optim object, in string format}

\item{elbo}{a pyro.inference object, in string format}

\item{inf_type}{currently just "SVI" is supported}

\item{steps}{number of step of the gradient-based VI algorithm}

\item{lr}{learning rate of the gradient-based VI algorithm}

\item{param_list}{named list of parameters for the specific model}

\item{MAP}{perform MAP or full bayesian inference (MAP is suggested as it converges better)}

\item{seed}{seed for torch random number generator}

\item{mixture}{initial mixture weights, if NULL then they are equal to 1/K}

\item{method}{informtion criteria to score the best solution (currently supports "AIC", "BIC" and "ICL)}
}
\value{
A congas (described in [\fun{init}]) object with following elements:
- *inference* :
  - *models* with all the runs performed
  - *model_selection* with information about the IC and run statistics
}
\description{
Selects the best number of clusters \code{k} to fit a dataset.
Options allow to select the type of model available in the package, parameters
for the Pyro optimizer and a number of different configurations to learn
Bayesian posteriors or MAP estimates.
}
\details{
While this function has a lot of personalizable arguments, the default
ones are usually good for most of the applications.
}
