% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cluster.R
\name{best_cluster}
\alias{best_cluster}
\title{Model selection with CONGAS.}
\usage{
best_cluster(
  X,
  model,
  clusters,
  optim = "ClippedAdam",
  elbo = "TraceEnum_ELBO",
  inf_type = "SVI",
  steps = 300,
  lr = 0.05,
  param_list = list(),
  MAP = TRUE,
  seed = 3,
  mixture = NULL,
  method = "AIC",
  norm_factor = NULL
)
}
\arguments{
\item{X}{}

\item{model}{Any of the model keywords available via \code{list_models()}.}

\item{clusters}{A vector of values for \code{k}.}

\item{optim}{}

\item{elbo}{}

\item{inf_type}{}

\item{steps}{}

\item{lr}{}

\item{param_list}{}

\item{MAP}{}

\item{seed}{}

\item{mixture}{}

\item{method}{}

\item{posteriors}{}

\item{step_post}{}
}
\value{

}
\description{
Selects the best number of clusters \code{k} to fit a dataset.
Options allow to select the type of model available in the package, parameters
for the Pyro optimizer and a number of different configurations to learn
Bayesian posteriors or MAP estimates.
}
