% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_object}
\alias{example_object}
\title{Example object from (R)Congas+}
\format{
Example object.
}
\usage{
data(example_object)
}
\description{
Example object from (R)Congas+
}
\examples{
data(example_object)
print(example_object)
}
\keyword{datasets}
