% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clusters.R
\name{filter_clusters}
\alias{filter_clusters}
\title{Filter output clusters}
\usage{
filter_clusters(x, ncells = 10, abundance = 0.03, remove = F)
}
\arguments{
\item{remove}{}
}
\value{

}
\description{
From a CONGAS fit object remove clusters that do not pass filters.
These include number of cells and abundance of cluster (mixing proportions).
}
\examples{

x = Rcongas::congas_example

print(x)

# Equivalent filters for this model
x \%>\% Rcongas:::filter_clusters(abundance = .5)

x \%>\% Rcongas:::filter_clusters(ncells = 150)

x \%>\% Rcongas:::filter_clusters(ncells = 150, remove = TRUE)
}
