% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_known_genes}
\alias{filter_known_genes}
\title{Filter genes from an RNA tibble.}
\usage{
filter_known_genes(x, what = "rm", specials = "MALAT1")
}
\arguments{
\item{x}{The tibble for RNA.}

\item{what}{A code string. If this contains "r" ribosomal genes are removed.
If this contains "m" mitocondrial genes are removed. Both are matched by
a regular expression.}

\item{specials}{A list of genes to remove a priori.}
}
\value{
The RNA tibble without the required gene entries.
}
\description{
Filter genes from an RNA tibble.
}
\examples{
data('example_input')
filter_known_genes(example_input$rna)
}
