% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_segments_and_return_cnv_table.R
\name{filter_segments_and_return_cnv_table}
\alias{filter_segments_and_return_cnv_table}
\title{Title}
\usage{
filter_segments_and_return_cnv_table(x, mu = 15, length = 2e+07)
}
\arguments{
\item{x}{}

\item{mu}{}

\item{length}{}
}
\value{

}
\description{
Title
}
