% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_congas.R
\name{fit_congas}
\alias{fit_congas}
\title{Fit an (R)CONGAS+ model}
\usage{
fit_congas(
  x,
  K,
  model_parameters,
  learning_rate = 0.05,
  latent_variables = "D",
  compile = FALSE,
  steps = 500,
  samples = 1,
  parallel = FALSE,
  model_selection = "ICL",
  temperature = 10
)
}
\arguments{
\item{x}{An \code{rcongasplus} object with the input dataset, constructed with \code{\link[Rcongas::init]{Rcongas::init}}.}

\item{K}{a vector of integers with the number of clusters we want to test}

\item{model_parameters}{a list with model hyperparameters. As errors coming from wrong hyperparameters initialization
are quite hard to troubleshoot is higly suggested to use \code{\link[Rcongas::auto_config_run]{Rcongas::auto_config_run()}} to generate
a template and eventually modify it.}

\item{learning_rate}{a learning rate for the Adam optimizer}

\item{latent_variables}{specify the nature of the latent variable modelling the copy number profile. Currently only "D" (discrete) is available}

\item{compile}{use JIT compiler for the Pyro ba}

\item{steps}{number of steps of optimization}

\item{samples}{Number of times a model is fit for each value of \code{K}.}

\item{model_selection}{information criteria to which perform the model selection (one of ICL, NLL, BIC, AIC)}
}
\value{
An object ot class \code{rcongasplus} with a slot \code{bset_fit} with the learned parameters for the selected model in tiblle format. A slot \code{runs}
with all the runs performed ordered by the selectde IC and a slot \code{model_selection} with all the information to perform model selection.
}
\description{
This function is a general interface for fitting a \href{https://github.com/Militeee/congas}{congas} Python model in R. The model briefly consist in a joint
mixture model over two modalities, currently scATAC and scRNA-seq. For more information about the theoretical fundations of the approach refer
to the vignette. This function performs modele selection over a specified number of clusters, using a specific information criterium (IC). ICs
and results for all the runs are, however, reported in the object.

The functions assume a list of model hyperparameters. As the the model formulation isquite complex, and those hyperparameters are extremely difficult to
set by hand we suggest the usage of the function \code{\link[Rcongas::auto_config_run]{Rcongas::auto_config_run()}}
}
\examples{
library(Rcongas)
\dontrun{
K <-  1:4
hyperparams <- auto_config_run(example_object, 1:4)

fit <- fit_congas(example_object, K = 1:4,learning_rate = 0.05, model_parameters = hyperparams)
}
}
