% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_clustering_results.R
\name{get_clones_ploidy}
\alias{get_clones_ploidy}
\title{Title}
\usage{
get_clones_ploidy(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  clusters = NULL
)
}
\arguments{
\item{chromosomes}{}
}
\value{

}
\description{
Title
}
