% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{Return expression level per segment}
\usage{
get_counts(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  normalise = TRUE,
  z_score = FALSE,
  sum_denominator = TRUE
)
}
\arguments{
\item{chromosomes}{Chromosome to filter}

\item{normalise}{should the counts be divided by the library size?}

\item{z_score}{returns the Z score for each segment instead of the actual expression value}

\item{inf_obj}{an Rcongas object}

\item{should}{the the denominator of the likelihood be included? (usually TRUE unless you are using very old versions)}
}
\value{
A tibble with expression values for each segment and cluster
}
\description{
Return expression level per segment
}
\examples{

library(Rcongas)

data('congas_example', package = 'Rcongas')

get_counts(congas_example)


}
