% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_extra_analysis.R
\name{get_DE_table}
\alias{get_DE_table}
\title{Title}
\usage{
get_DE_table(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  cut_pvalue = 0.001,
  cut_lfc = 0.25
)
}
\arguments{
\item{cut_lfc}{}
}
\value{

}
\description{
Title
}
