% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit.R
\name{get_fit}
\alias{get_fit}
\title{Get fit data.}
\usage{
get_fit(x, what = "segment_factors")
}
\arguments{
\item{x}{}

\item{what}{Any of \code{"segment_factors"}, \code{"CNA"},  \code{"posterior_CNA"},
\code{"mixing_proportions"},  \code{"cluster_assignments"},  or \code{"z_nk"}.}
}
\value{
A tibble.
}
\description{
General fit accessing getter function to return
segment_factors, CNAs (point estimates), posterior CNA (distribution),
mixing proportions, clustering assignments or the latent responsibilities.
}
