% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input.R
\name{get_input}
\alias{get_input}
\title{Get input data.}
\usage{
get_input(x, what = "data")
}
\arguments{
\item{what}{Any of \code{"data"}, \code{"segmentation"} or
\code{"normalisation"}.}
}
\value{
A tibble.
}
\description{
General data accessing getter function to return
input data, segmentation or normalisation.
}
