% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_raw_data.R
\name{get_input_raw_data}
\alias{get_input_raw_data}
\title{Title}
\usage{
get_input_raw_data(
  x,
  transpose = FALSE,
  only_clustered_cells = TRUE,
  as_tibble = FALSE
)
}
\arguments{
\item{transpose}{}

\item{as_tibble}{}

\item{all_cells}{}
}
\value{

}
\description{
Title
}
