% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_input_data.R
\name{get_input_segmentation}
\alias{get_input_segmentation}
\title{Title}
\usage{
get_input_segmentation(x, chromosomes = paste0("chr", c(1:22, "X", "Y")))
}
\arguments{
\item{chromosomes}{}
}
\value{

}
\description{
Title
}
