% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segment_test_counts.R
\name{get_segment_test_counts}
\alias{get_segment_test_counts}
\title{Wilcoxon test for difference in RNA counts among groups.}
\usage{
get_segment_test_counts(x, group1, group2, cutoff_p = 0.01)
}
\arguments{
\item{x}{Input object with clusters.}

\item{group1}{Group of cluster ids (vs \code{group2}).}

\item{group2}{Group of cluster ids (vs \code{group1}).}

\item{cutoff_p}{P-value cutoff}
}
\value{
A tibble with segment coordinates, p-values and a \code{sign} status
for the significance of the test.
}
\description{
Performs a \code{wilcox.test} for the counts in each segments,
split by pair of groups. A group can be a list of cluster ids; the test
p-value can be passed as a parameter, and is adjusted for Family-Wise Error
Rate (FWER) via Bonferroni correction.
}
\examples{

x = Rcongas::congas_example

# Compare groups
get_segment_test_counts(x, "c1", "c2")

# More stringent test
get_segment_test_counts(x, "c1", "c2", cutoff_p = 1e-4)
}
