% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Title}
\usage{
init(
  data,
  cnv_data,
  chromosomes = paste0("chr", (1:22)),
  fun = sum,
  online = FALSE,
  reference_genome = "hg38",
  correct_bins = TRUE,
  gene_format = "hgnc_symbol",
  description = "My CONGAS dataset"
)
}
\arguments{
\item{data}{}

\item{cnv_data}{}

\item{chromosomes}{}

\item{fun}{}

\item{online}{}

\item{reference_genome}{}

\item{correct_bins}{}

\item{gene_format}{}

\item{description}{}
}
\value{

}
\description{
Title
}
