% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a CONGAS object}
\usage{
init(
  data,
  cnv_data,
  chromosomes = paste0("chr", (1:22)),
  fun = sum,
  online = FALSE,
  reference_genome = "hg38",
  correct_bins = TRUE,
  gene_format = "hgnc_symbol",
  description = "My CONGAS dataset"
)
}
\arguments{
\item{data}{}

\item{cnv_data}{cnv table, a data.frame with columns c("chr", "from", "to", "tot"). 
Where tot is the estimated bulk copy number value}

\item{chromosomes}{which chromosome to include in the analysis}

\item{fun}{a valid R function to aggregate the data: should be defined as R^n -> R}

\item{online}{get gene position information by querying BioMart}

\item{reference_genome}{reference assembly of human or mouse genome}

\item{correct_bins}{filter small segments, and then defragment them by merging
those with the same CNV}

\item{gene_format}{gene format from those available on biomaRt}

\item{description}{a description for the object}
}
\value{
Rcongas object
}
\description{
Generates an Rcongas object. Automatically maps genes to given CNV segments and calculate 
the sum of gene counts. It retrieves gene positions from BioMartR, if the packages is not installed we suggest to use `online=FALSE`
}
