% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Title}
\usage{
init(
  rna,
  atac,
  segmentation,
  normalisation_factors,
  rna_likelihood = "G",
  atac_likelihood = "NB",
  reference_genome = "GRCh38",
  description = "Your (R)CONGAS+ model"
)
}
\arguments{
\item{rna}{}

\item{atac}{}

\item{segmentation}{}

\item{normalisation_factors}{}

\item{rna_likelihood}{}

\item{atac_likelihood}{}

\item{reference_genome}{}

\item{description}{}
}
\value{

}
\description{
Title
}
