% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patel_gbmtn}
\alias{patel_gbmtn}
\title{Glioblastoma tumour and normal (SmartSeq RNA, z-score)}
\format{
(R)CONGAS+ object.
}
\usage{
data(patel_gbmtn)
}
\description{
@description (R)CONGAS+ object for the glioblastoma dataset,
first used in the original CONGAS paper. The paper providing these data
is Patel, Anoop P., et al. "Single-cell RNA-seq highlights intratumoral
heterogeneity in primary glioblastoma." Science 344.6190 (2014): 1396-1401.
}
\examples{
data(patel_gbmtn)
print(patel_gbmtn)
}
\keyword{datasets}
