% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongas.R
\name{plot.rcongas}
\alias{plot.rcongas}
\title{Print for an object of class \code{'rcongas'}.}
\usage{
\method{plot}{rcongas}(x, ...)
}
\arguments{
\item{...}{Extra parameters.}
}
\value{
A ggplot object for the plot.
}
\description{
Function \code{plot_gw_cna_profiles} is used to plot
the object.
}
\examples{

x = Rcongas::congas_example

plot(x)
}
