% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts_per_segment.R
\name{plot_counts_per_segment}
\alias{plot_counts_per_segment}
\title{Title}
\usage{
plot_counts_per_segment(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  annotate_from = 3,
  ...
)
}
\arguments{
\item{chromosomes}{}
}
\value{

}
\description{
Title
}
