% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts_rna_segments.R
\name{plot_counts_rna_segments}
\alias{plot_counts_rna_segments}
\title{Title}
\usage{
plot_counts_rna_segments(
  x,
  normalised = TRUE,
  z_score = FALSE,
  z_score_cutoff = 5,
  sum_denominator = FALSE,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  ...
)
}
\arguments{
\item{z_score}{}

\item{z_score_cutoff}{Values above or below this cut are capped and reported
in the plot caption.}

\item{sum_denominator}{}

\item{chromosomes}{}

\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

x = Rcongas::congas_example

print(x)

# Shows all data available
plot_counts_rna_segments(x)

# Not normalised
plot_counts_rna_segments(x, normalised = FALSE)

# normalised and z-score
plot_counts_rna_segments(x, normalised = TRUE, z_score = TRUE) 
}
