% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input_rna_segments.R
\name{plot_counts_rna_segments}
\alias{plot_counts_rna_segments}
\title{Title}
\usage{
plot_counts_rna_segments(x, normalised = TRUE)
}
\arguments{
\item{normalised}{}
}
\value{

}
\description{
Title
}
\examples{

input = get_counts_matrix(x)
}
