% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot data.}
\usage{
plot_data(
  x,
  segments = get_input(x, what = "segmentation") \%>\% pull(segment_id),
  what = "histogram"
)
}
\arguments{
\item{segments}{Can subset the plot only to certain segments, which are indexed
by the \code{segment_id} format in the tibble returned by function \code{get_input}
with \code{what = 'segmentation'} as parameter.}

\item{what}{Any of \code{"segment_factors"}, \code{"CNA"},  \code{"posterior_CNA"},
\code{"mixing_proportions"},  \code{"cluster_assignments"},  or \code{"z_nk"}.}
}
\value{
A tibble.
}
\description{
General plot for data, accessing getter function to return
segment_factors, CNAs (point estimates), posterior CNA (distribution),
mixing proportions, clustering assignments or the latent responsibilities.
}
