% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DE_gw.R
\name{plot_DE_gw}
\alias{plot_DE_gw}
\title{Title}
\usage{
plot_DE_gw(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  cut_pvalue = 0.001,
  cut_lfc = 0.25
)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
\examples{

x = Rcongas::congas_example

print(x)

# Shows all data available
plot_DE_gw(x)

# More stringent parameters
plot_DE_gw(x, cut_pvalue = 1e-5)

# More stringent parameters
plot_DE_gw(x, cut_lfc = 2)
}
