% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot fits}
\usage{
plot_fit(x, what = "CNA", ...)
}
\arguments{
\item{what}{Any of \code{"histogram"},  \code{"lineplot"},  or \code{"heatmap"}.}

\item{segments}{Can subset the plot only to certain segments, which are indexed
by the \code{segment_id} format in the tibble returned by function \code{get_input}
with \code{what = 'segmentation'} as parameter.}
}
\value{
A \code{ggplot} or \code{cowplot} figure, depending on the number of
modalities and plot required.
}
\description{
General plot for fits, showing them as .....
}
