% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot fits.}
\usage{
plot_fit(x, what = "CNA", ...)
}
\arguments{
\item{x}{An object of class \code{rcongasplus}.}

\item{what}{Any of \code{"CNA"},  \code{"density"}, \code{"mixing_proportions"}, \code{"heatmap"}
or \code{"scores"}.}

\item{...}{Parameters forwarded to the internal plotting functions.}
}
\value{
A \code{ggplot} plot, or a more complex \code{cowplot} figure.
}
\description{
This function is like \code{\link{plot_data}}, but shows information about a fit model.

As \code{\link{plot_data}}, this function uses a \code{what} parameter
to dispatch visualization to a number of internal functions. For the fits one can visualize:
\itemize{
\item (\code{what = "CNA"}) A genome-wide plot of Copy Number Alteration profiles (inferred) per cluster;
\item (\code{what = "mixing_proportions"}) The normalized size of each cluster, per modality;
\item (\code{what = "density"}) The density per cluster and segment, split by modality. By default this is shown
only for segments that differ for a segment CNA value among one of the inferred clusters;
\item (\code{what = "heatmap"}) The same heatmap plot from \code{\link{plot_data}} with
\code{what = "heatmap"}, where rows are sorted by cluster and cluster annotations reported;
\item (\code{what = "scores"}) The scores used for model selection;
\item (\code{what = "posterior_CNA"}) The posterior probability for CNA values;
}

This function has the same logic of \code{\link{plot_data}} with respect to the ellipsis and
input parameters.
}
\examples{
data("example_object")

# Genome-wide segments plo
plot_fit(example_object, what = 'CNA')

# Density plot
plot_fit(example_object, what = 'density')

# Mixing proportions
plot_fit(example_object, what = 'mixing_proportions')

# Fit heatmap
plot_fit(example_object, what = 'heatmap')

# Scores for model selection
plot_fit(example_object, what = 'scores')

# Posterior for CNAs
plot_fit(example_object, what = 'posterior_CNA')
}
