% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gw_cna_profiles.R
\name{plot_gw_cna_profiles}
\alias{plot_gw_cna_profiles}
\title{Title}
\usage{
plot_gw_cna_profiles(
  x,
  whole_genome = FALSE,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  ...
)
}
\arguments{
\item{chromosomes}{}
}
\value{

}
\description{
Title
}
\examples{
x = Rcongas::congas_example

# Default view 
plot_gw_cna_profiles(x)

# Avoid facet (shows real chromosome width)
plot_gw_cna_profiles(x, whole_genome = TRUE)

# Subset to one chromosome
plot_gw_cna_profiles(x, chromosomes = c('chr1'))

# Forward alpha to get_clusters_ploidy (changes what is highlighted)
plot_gw_cna_profiles(x, whole_genome = TRUE, alpha = 0.001)
}
