% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_variables.R
\name{plot_latent_variables}
\alias{plot_latent_variables}
\title{Plot latent varibles.}
\usage{
plot_latent_variables(x, cutoff_assignment = 0)
}
\arguments{
\item{x}{Input object with clusters}

\item{cutoff_assignment}{A value in $[0, 1)$ so that values below this are
set to \code{NA}}
}
\value{
A ggplot object
}
\description{
Plots the clustering responsibilities (posterior over $z$)
for the available clusters. It can put \code{NA} to all entries where the hard 
clustering assignments - max of the posterior - is below a cutoff.
}
\examples{
x = Rcongas::congas_example

# Default (cutoff 0)
plot_latent_variables(x) 

# >50\% assignment probability
plot_latent_variables(x, cutoff_assignment = .5) 

# >90\% assignment probability
plot_latent_variables(x, cutoff_assignment = .9) 

# >97.5\% assignment probability
plot_latent_variables(x, cutoff_assignment = .975) 
}
