% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_variables.R
\name{plot_latent_variables}
\alias{plot_latent_variables}
\title{Title}
\usage{
plot_latent_variables(x, cutoff_assignment = 0)
}
\arguments{
\item{cutoff_assignment}{}
}
\value{

}
\description{
Title
}
\examples{
x = Rcongas::congas_example

# Default (cutoff 0)
plot_latent_variables(x) 

# >50\% assignment probability
plot_latent_variables(x, cutoff_assignment = .5) 

# >90\% assignment probability
plot_latent_variables(x, cutoff_assignment = .9) 

# >97.5\% assignment probability
plot_latent_variables(x, cutoff_assignment = .975) 
}
