% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongas.R
\name{print.rcongas}
\alias{print.rcongas}
\title{Print for an object of class \code{'rcongas'}.}
\usage{
\method{print}{rcongas}(x, ...)
}
\arguments{
\item{x}{Object of class \code{'rcongas'}.}

\item{...}{Extra parameters.}
}
\value{
Nothing.
}
\description{
Print for an object of class \code{'rcongas'}.
}
\examples{

x = Rcongas::congas_example

print(x)
}
