% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongasplus.R
\name{print.rcongasplus}
\alias{print.rcongasplus}
\title{Print for an object of class \code{'rcongasplus'}.}
\usage{
print.rcongasplus(x, ...)
}
\arguments{
\item{x}{An object of class \code{'rcongasplus'}.}

\item{...}{Unused ellipsis.}
}
\value{
Nothing.
}
\description{
Print for an object of class \code{'rcongasplus'}.
}
\examples{
data("example_object")
print(example_object)
}
