% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{run_inference}
\alias{run_inference}
\title{Title}
\usage{
run_inference(
  X,
  model,
  optim = "ClippedAdam",
  elbo = "TraceEnum_ELBO",
  inf_type = "SVI",
  steps = 300,
  lr = 0.05,
  param_list = list(),
  MAP = TRUE,
  posteriors = FALSE,
  seed = 3,
  step_post = 300,
  verbose = FALSE,
  filt_merge = 0.5
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
