% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{run_inference}
\alias{run_inference}
\title{Run CONGAS inference}
\usage{
run_inference(
  X,
  model,
  optim = "ClippedAdam",
  elbo = "TraceEnum_ELBO",
  inf_type = "SVI",
  steps = 300,
  lr = 0.05,
  param_list = list(),
  MAP = TRUE,
  seed = 3,
  rerun = F
)
}
\arguments{
\item{step_post}{}
}
\value{

}
\description{
This function acts as the main interface with the Python package. It allows to tweak
most of the parameters of the Python model. Most of default is good for most of the applications
}
