% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_rcongas.R
\name{summary.rcongas}
\alias{summary.rcongas}
\title{Summary for an object of class \code{'rcongas'} is a print.}
\usage{
\method{summary}{rcongas}(object, ...)
}
\arguments{
\item{object}{An obj of class \code{'rcongas'}.}

\item{...}{}
}
\value{
See \code{\link{print}}.
}
\description{
Summary for an object of class \code{'rcongas'} is a print.
}
\examples{
data(fit_example)
summary(fit_example$best)
}
